+++ Sources/SwiftUIKit/Sheets/SheetContext.swift
@@ -20,10 +20,15 @@ import SwiftUI
  ```swift
  @ObservedObject var context = SheetContext()
  
+ view.sheet(context: context)   // or, for more flexibility:
  view.sheet(isPresented: $context.isActive, content: context.sheet)
  
  context.present(Text("Hello, world!"))
  ```
+ 
+ The context-specific `sheet` modifier is more conveinent if
+ you use contexts, but the binding/content-specific one is a
+ bit more flexible and can be used in other scenarios.
  
  You can also use the `SheetProvider` protocol to present an
  alert for basically anything, e.g. an enum with cases, that
@@ -34,6 +39,12 @@ public class SheetContext: ObservableObject {
     public init() {}
     
     @Published public var isActive = false
+    
+    public var isActiveBinding: Binding<Bool> {
+        .init(get: { self.isActive },
+              set: { self.isActive = $0 }
+        )
+    }
     
     public private(set) var sheetView: AnyView? {
         didSet { isActive = sheetView != nil }