+++ Sources/SwiftUIKit/Toasts/ToastContext.swift
@@ -19,10 +19,18 @@ import SwiftUI
  ```swift
  @ObservedObject var context = ToastContext()
  
+ view.toast(context: context)   // or, for more flexibility:
  view.toast(isPresented: $context.isActive, content: context.toast)
  
  context.present(Text("Hello, world!").background(Color.white))
  ```
+ 
+ The context-specific `toast` modifier is more conveinent if
+ you use contexts, but the binding/content-specific one is a
+ bit more flexible and can be used in other scenarios.
+ 
+ The `toast` modifier also lets you specify a `duration` and
+ a `style` as well.
  
  You can also use the `ToastProvider` protocol, to present a
  toast for basically anything, e.g. an enum with cases, that
@@ -33,6 +41,12 @@ public class ToastContext: ObservableObject {
     public init() {}
     
     @Published public var isActive = false
+    
+    public var isActiveBinding: Binding<Bool> {
+        .init(get: { self.isActive },
+              set: { self.isActive = $0 }
+        )
+    }
     
     public private(set) var toastView: AnyView? {
         didSet { isActive = toastView != nil }